Create BaseImageGenerationTask
